feat(auth): refactor authentication service and types
