+++ app/types/auth.type.ts
